feat: add basic asset wrapper
This commit adds an initial implementation of the asset wrapping
protocol which allows users to create bundles of assets that can be used
as collateral in the pawn finance protocol. The implementation currently
only supports ERC20 tokens, but will be extended to include ERC721 and
ERC1155 tokens in the future.