feat(core): add loan core
This commit adds an implementation of the LoanCore contract. This
implements the core functionality of the PawnFi lending platform
It also adds a full test suite of the LoanCore contract.

There is some functionality that still remains after this PR. The TODO
list can be found in a block comment at the top of LoanCore.sol, as well
as in inline comments throughout