feat: add fee claimer role
This commit adds a new admin role to LoanCore for claiming fees.
I imagine there may be a situation where the admin who controls
upgrades, etc. may not be the same as the admin who claims fees. For
example if we institute a decentralized governance mechanism we will
want them to be able to handle upgrades, but _not_ to claim the protocol
fees.