feat: add claim functionality to vault
This commit adds the claim functionality to the AssetVault.

This allows someone to call an arbitrary function from the
vault's context (i.e. claim(token)).

The user who can call functions on behalf if the vault is one of the
following:
- the current owner of the contract
- an address that the current owner of the contract explicitly delegates
call ability to (i.e. LoanCore delegates it to the borrower of the
relevant loan).

It uses a whitelist contract to protect the lender and any
system holding a vault NFT as collateral from
having the underlying collateral sneakily removed.

The whitelist contract is a singleton admin contract which has two main
components:
- a dynamic whitelist where the admin can add (contract address,
function selector) pairs that are allowed to be called
- a global blacklist of functions that can never be called in any
context (i.e. token transfers for removal from the vault)