feat: add vault contracts (#56)
* feat: add vault contracts

This commit adds contracts and tests for a segregated vault system for
asset wrappers. The main contracts are as follows:
- VaultFactory - the NFT contract which spawns new vaults upon
`initializeBundle` calls. It mints a new NFT token representing
ownership of the vault and sends it to the creator
- AssetVault - Contract for storing segregated bundles of assets,
including ERC20, ERC721, ERC1155, and ETH. Users can create a vault from
the factory and safely send in their assets to the vault, using it as
collateral in bundle form. When they want to withdraw, they must
explicity `enableWithdraw` on their vault, which changes the state and
allows withdrawals of any deposited assets. Note that once withdrawals
are enabled they cannot be disabled - this is a safety feature for
integrations like the OriginationController, which need to assert that
the assets a potential lender saw in the vault are still there.

* feat: add claim functionality to vault (#58)

This commit adds the claim functionality to the AssetVault.

This allows someone to call an arbitrary function from the
vault's context (i.e. claim(token)).

The user who can call functions on behalf if the vault is one of the
following:
- the current owner of the contract
- an address that the current owner of the contract explicitly delegates
call ability to (i.e. LoanCore delegates it to the borrower of the
relevant loan).

It uses a whitelist contract to protect the lender and any
system holding a vault NFT as collateral from
having the underlying collateral sneakily removed.

The whitelist contract is a singleton admin contract which has two main
components:
- a dynamic whitelist where the admin can add (contract address,
function selector) pairs that are allowed to be called
- a global blacklist of functions that can never be called in any
context (i.e. token transfers for removal from the vault)